rename useLocalStorage file
