+++ src/hooks/index.ts
@@ -1,3 +1,3 @@
 export { useRecognition } from './speechAPI';
-export { default as useLocalStorage } from './storageAPI/useLocalStorage';
+export { default as useLocalStorage } from './storageAPI';
 export { useBreakPoint } from './useBreakPoint';
